requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
openpyxl==3.1.2
chromadb==0.4.18
sentence-transformers==2.7.0
torch==2.1.1
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
openai==0.28.1
sqlmodel==0.0.22
jinja2==3.1.6
# Optional: uncomment for Neo4j support
# neo4j==5.14.0
